# server/docs/swagger.yaml
basePath: /
definitions:
...
  community.CircleRequest:
    properties:
      backImage:
        description: 圈子背景图
        type: string
      checkStatus:
        description: 审核状态：0 未处理 1 通过，2驳回
        type: integer
      circleClassifyId:
        description: 圈子分类_编号
        type: integer
      createdAt:
        description: 创建时间
        type: string
      des:
        description: 圈子简介
        type: string
      id:
        description: 主键ID
        type: integer
      logo:
        description: 圈子Logo
        type: string
      name:
        description: 圈子名称
        type: string
      protocol:
        description: 圈子规约
        type: string
      slogan:
        description: 圈子标语
        type: string
      type:
        description: 类型：0官方圈子 ，1用户圈子
        type: integer
    type: object
...
  request.CreateCircleRequestReq:
    properties:
      backImage:
        description: 圈子背景图
        type: string
      circleClassifyId:
        description: 圈子分类_编号
        type: integer
      des:
        description: 圈子简介
        type: string
      logo:
        description: 圈子Logo
        type: string
      name:
        description: 圈子名称
        type: string
      protocol:
        description: 圈子规约
        type: string
      slogan:
        description: 圈子标语
        type: string
    type: object
...
  request.DeleteCircleUserReq:
    properties:
      circleId:
        description: 圈子_编号
        type: integer
      userId:
        description: 用户ID
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.SetUserCurCircleReq:
    properties:
      circleId:
        description: 圈子_编号
        type: integer
    type: object
...
  response.DataResult:
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /app/circle/createCircleRequest:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建CircleRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateCircleRequestReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建CircleRequest
      tags:
      - App_Circle
  /app/circle/deleteCircleUser:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除CircleUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteCircleUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除CircleUser
      tags:
      - App_Circle
  /app/circle/deleteCircleUserByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除CircleUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除CircleUser
      tags:
      - App_Circle
  /app/circle/findCircle:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.Circle
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/community.Circle'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询Circle
      tags:
      - App_Circle
  /app/circle/findCircleRequest:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleUser
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/community.CircleRequest'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询CircleRequest
      tags:
      - App_Circle
  /app/circle/findCircleUser:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleUserInfo
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/community.CircleUserInfo'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询CircleUserInfo
      tags:
      - App_Circle
  /app/circle/getCircleClassifyList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 描述
        in: query
        name: des
        type: string
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleClassify
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.CircleClassify'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取CircleClassify列表
      tags:
      - App_Circle
  /app/circle/getCircleClassifyListAll:
    get:
      consumes:
      - application/json
      parameters:
      - description: 描述
        in: query
        name: des
        type: string
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleClassify
          schema:
            allOf:
            - $ref: '#/definitions/response.DataResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/community.CircleClassify'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取CircleClassify列表
      tags:
      - App_Circle
  /app/circle/getCircleForumPostsList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 类别（0视频、1动态、2资讯、3公告、4文章、5问答、6建议）
        in: query
        name: category
        type: integer
      - description: 圈子_编号
        in: query
        name: circleId
        type: integer
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 帖子分类编号
        in: query
        name: groupId
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 精华(0否 1是)
        in: query
        name: marrow
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: 置顶(0否 1是)
        in: query
        name: top
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumPostsBaseInfo
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumPostsBaseInfo'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取圈子ForumPosts列表
      tags:
      - App_Circle
  /app/circle/getCircleList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 搜索名字：圈子名称或圈子简介
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 类型：0官方圈子、1用户圈子、2小区
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleBaseInfo
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.CircleBaseInfo'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取Circle列表
      tags:
      - App_Circle
  /app/circle/getCircleRequestList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 审核状态：0 未处理 1 通过，2驳回
        in: query
        name: checkStatus
        type: integer
      - description: 圈子分类_编号
        in: query
        name: circleClassifyId
        type: integer
      - description: 圈子简介
        in: query
        name: des
        type: string
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 圈子名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 圈子标语
        in: query
        name: slogan
        type: string
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 类型：0官方圈子 ，1用户圈子
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleRequest
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.CircleRequest'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取CircleRequest列表
      tags:
      - App_Circle
  /app/circle/getCircleUserList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 圈子_编号
        in: query
        name: circleId
        type: integer
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleUserInfo
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.CircleUserInfo'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取CircleUserInfo列表
      tags:
      - App_Circle
  /app/circle/getSelfCircleList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 搜索名字：圈子名称或圈子简介
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 类型：0官方圈子、1用户圈子、2小区
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.CircleBaseInfo
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.CircleBaseInfo'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取用户加入的Circle列表
      tags:
      - App_Circle
  /app/circle/getUserCircleForumPostsList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 类别（0视频、1动态、2资讯、3公告、4文章、5问答、6建议）
        in: query
        name: category
        type: integer
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 帖子分类编号
        in: query
        name: groupId
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 精华(0否 1是)
        in: query
        name: marrow
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: 置顶(0否 1是)
        in: query
        name: top
        type: integer
      - description: 用户ID
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumPostsBaseInfo
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumPostsBaseInfo'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户加入圈子的所有动态列表
      tags:
      - App_Circle
  /app/circle/setUserCurCircle:
    post:
      consumes:
      - application/json
      parameters:
      - description: 设置用户当前圈子
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SetUserCurCircleReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"设置成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 设置用户当前圈子
      tags:
      - App_Circle
  /app/circle/updateCircle:
    put:
      consumes:
      - application/json
      parameters:
      - description: (圈子管理者)更新Circle
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCircleReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: (圈子管理者)更新Circle
      tags:
      - App_Circle
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"